fix: consolidate common logic for component transformation and dictionary loading
